+++ lib/solana_studio.rb
@@ -3,8 +3,10 @@ require_relative "solana/borsh"
 require_relative "solana/client"
 require_relative "solana/transaction"
 require_relative "solana/spl_token"
+require_relative "solana/system_program"
+require_relative "solana/nonce_account"
 require_relative "solana/auth_verifier"
 
 module SolanaStudio
-  VERSION = "0.4.5"
+  VERSION = "0.4.6"
 end